fix: Type for options in autodetected renderer was incorrectly inferring any passed type
Changed `renderHook` in `ReactHooksRenderer` type to not use a generic for options but rather the react specific `RendererOptions`.

Moved `ReactHooksRenderer` into react specific types as all auto-detected renders are react renderers.  All modules are currently exporting both the shared types and the react types so this a non-breaking change.
